@@ -1,7 +1,7 @@
 {
-  "name": "frontend",
+  "name": "personal-protfolio",
   "private": true,
-  "version": "0.0.0",
+  "version": "1.0.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
@@ -10,44 +10,44 @@
     "lint": "eslint . --ext .ts,.tsx"
   },
   "dependencies": {
+    "@react-hook/window-scroll": "^1.3.0",
+    "@sanity/client": "^4.0.1",
+    "@sanity/image-url": "^1.0.2",
+    "@tanstack/react-query": "^4.23.0",
+    "dayjs": "^1.11.7",
+    "framer-motion": "^8.5.3",
+    "lodash": "^4.17.21",
+    "node-sass": "^8.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "@react-hook/window-scroll": "^1.3.0",
-    "@sanity/client": "^3.1.0",
-    "@sanity/image-url": "^1.0.1",
-    "dayjs": "^1.10.7",
-    "framer-motion": "^6.2.6",
-    "lodash": "^4.17.21",
-    "node-sass": "^7.0.1",
-    "react-icons": "^4.3.1",
-    "react-intersection-observer": "^8.33.1",
-    "react-modal": "^3.14.4",
-    "react-router-dom": "^6.2.2",
-    "recoil": "^0.6.1",
-    "sass": "^1.49.8",
-    "three": "^0.138.0",
+    "react-icons": "^4.7.1",
+    "react-intersection-observer": "^9.4.1",
+    "react-modal": "^3.16.1",
+    "react-router-dom": "^6.8.0",
+    "sass": "^1.57.1",
     "url": "^0.11.0",
     "vanta": "^0.5.21"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.178",
-    "@types/node": "^17.0.18",
+    "@trivago/prettier-plugin-sort-imports": "^4.0.0",
+    "@types/lodash": "^4.14.191",
+    "@types/node": "^18.11.18",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
     "@types/react-modal": "^3.13.1",
-    "@types/three": "^0.137.0",
-    "postcss": "^8.4.7",
+    "@types/three": "^0.148.1",
+    "@typescript-eslint/eslint-plugin": "^5.49.0",
+    "@typescript-eslint/parser": "^5.49.0",
+    "@vitejs/plugin-react": "^3.0.1",
+    "eslint": "^8.32.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.1",
+    "postcss": "^8.4.21",
+    "prettier": "^2.8.3",
     "prettier-plugin-css-order": "^1.1.0",
-    "@types/react": "^18.0.17",
-    "@types/react-dom": "^18.0.6",
-    "@typescript-eslint/eslint-plugin": "^5.39.0",
-    "@typescript-eslint/parser": "^5.39.0",
-    "@vitejs/plugin-react": "^2.1.0",
-    "eslint": "^8.25.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.6.1",
-    "eslint-plugin-react": "^7.31.9",
-    "prettier": "^2.7.1",
-    "typescript": "^4.6.4",
-    "vite": "^3.1.0"
+    "typescript": "^4.9.4",
+    "vite": "^4.0.4"
   }
 }
