@@ -1,21 +1,24 @@
-import App from './App'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom'
-import { RecoilRoot } from 'recoil'
+
+import App from './app'
 import './main.scss'
-ReactDOM.render(
+
+const queryClient = new QueryClient()
+const container = document.getElementById('root') as HTMLElement
+const root = createRoot(container)
+
+root.render(
   <React.StrictMode>
-    {/* eslint-disable-next-line @typescript-eslint/ban-ts-comment */}
-    {/* @ts-ignore */}
-    <RecoilRoot>
+    <QueryClientProvider client={queryClient}>
       <BrowserRouter>
         <Routes>
           <Route path="/" element={<App />} />
           <Route path="*" element={<Navigate to="/" />} />
         </Routes>
       </BrowserRouter>
-    </RecoilRoot>
-  </React.StrictMode>,
-  document.getElementById('root')
+    </QueryClientProvider>
+  </React.StrictMode>
 )
